Технический патч с обновление RobustToolbox до 267.3.0 (#353)
* Великий фильтр

* Перевыпуск болезней p.1

* Update submodule to 262.0.0 (#38174)

* Update submodule to 263.0.0 (#38524)

* Update submodule to 264.0.0 (#38629)

* Add FixedPoint2TypeParser (#38169)

* Update submodule to v265.0.0

* Content fixes for engine PR (#39086)

* content changes for engine PR

* rerun tests

* Update content to remove redundant Robust math types (#38930)

Vector3, Vector4, Matrix4, and Quaternion are now gone. This switches us over to System.Numerics where applicable.

* Content fix for `LocalizedCommands` `Loc` change (#38133)

* Content fix for LocalizedCommands Loc change

* Switch to LocalizationManager

* wizden robust up fixes

* Re-id 'Medical Doctor' guide entry to 'MedicalDoctor' (#39029)

chore: re-id 'Medical Doctor' guide entry

* naming patch wizden

* Update RT to 266.0.0

* Move scale command to content and turn it into a toolshed command (#39349)

* scale command

* fix namespaces

* fix DoAfter DistanceThreshold (#39276)

* fix: auto-update mailing unit + gas canister UIs on state (#39289)

* fix: auto-update mailing unit + gas canister UIs

* fix: make FollowerComponent auto handle state

* refactor: kill AfterAutoHandleState for Follower

* flakeops

* NF conflict Wizden

* Update engine to v267.1.0 (#40445)

* Update OpenTK to latest (#39227)


* Update engine to v267.2.0 (#40560)

* wizden fix rsi

* Update engine to v267.3.0 (#40899)

* version 267.3.0

---------

+++ Content.Server/_Backmen/Disease/DiseaseDiagnosisSystem.cs
@@ -116,7 +116,8 @@ public sealed class DiseaseDiagnosisSystem : EntitySystem
     /// </summary>
     private void OnAfterInteractUsing(EntityUid uid, DiseaseDiagnoserComponent component, AfterInteractUsingEvent args)
     {
-        var machine = Comp<DiseaseMachineComponent>(uid);
+        if (!TryComp<DiseaseMachineComponent>(uid, out var machine))
+            return;
         if (args.Handled || !args.CanReach)
             return;
 
@@ -163,9 +164,12 @@ public sealed class DiseaseDiagnosisSystem : EntitySystem
             _popupSystem.PopupEntity(Loc.GetString("diagnoser-cant-use-swab", ("machine", uid), ("swab", args.Used)), uid, args.User);
             return;
         }
+        
+        if (!TryComp<DiseaseMachineComponent>(uid, out var vaccineMachine))
+            return;
+            
         _popupSystem.PopupEntity(Loc.GetString("machine-insert-item", ("machine", uid), ("item", args.Used), ("user", args.User)), uid, args.User);
-        var machine = Comp<DiseaseMachineComponent>(uid);
-        machine.Disease = swab.Disease;
+        vaccineMachine.Disease = swab.Disease;
         EntityManager.DeleteEntity(args.Used);
 
         EnsureComp<DiseaseMachineRunningComponent>(uid);
@@ -212,6 +216,16 @@ public sealed class DiseaseDiagnosisSystem : EntitySystem
             report.AddMarkup(Loc.GetString("diagnoser-disease-report-not-infectious"));
             report.PushNewline();
         }
+        string cureResistLine = string.Empty;
+        cureResistLine += disease.CureResist switch
+        {
+            < 0f => Loc.GetString("diagnoser-disease-report-cureresist-none"),
+            <= 0.05f => Loc.GetString("diagnoser-disease-report-cureresist-low"),
+            <= 0.14f => Loc.GetString("diagnoser-disease-report-cureresist-medium"),
+            _ => Loc.GetString("diagnoser-disease-report-cureresist-high")
+        };
+        report.AddMarkup(cureResistLine);
+        report.PushNewline();
 
         // Add Cures
         if (disease.Cures.Count == 0)
